Move LocalCalendar, LocalEvent to separate package
